+++ server/src/test/java/io/unitycatalog/server/base/BaseCRUDTestWithMockCredentials.java
@@ -1,10 +1,14 @@
 package io.unitycatalog.server.base;
 
-import static org.mockito.ArgumentMatchers.any;
+import static org.assertj.core.api.Assertions.*;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
+import static org.mockito.ArgumentMatchers.argThat;
 import static org.mockito.Mockito.doReturn;
 import static org.mockito.Mockito.mock;
 
 import com.google.auth.oauth2.AccessToken;
+import io.unitycatalog.client.model.TemporaryCredentials;
+import io.unitycatalog.server.exception.BaseException;
 import io.unitycatalog.server.service.credential.CredentialOperations;
 import io.unitycatalog.server.service.credential.aws.AwsCredentialVendor;
 import io.unitycatalog.server.service.credential.azure.AzureCredential;
@@ -12,7 +16,13 @@ import io.unitycatalog.server.service.credential.azure.AzureCredentialVendor;
 import io.unitycatalog.server.service.credential.gcp.GcpCredentialVendor;
 import java.time.Instant;
 import java.util.Date;
+import java.util.List;
+import java.util.concurrent.CompletionException;
+import java.util.stream.Stream;
+
+import org.junit.jupiter.params.provider.Arguments;
 import software.amazon.awssdk.services.sts.model.Credentials;
+import software.amazon.awssdk.services.sts.model.StsException;
 
 public abstract class BaseCRUDTestWithMockCredentials extends BaseCRUDTest {
   @Override
@@ -41,7 +51,17 @@ public abstract class BaseCRUDTestWithMockCredentials extends BaseCRUDTest {
             .secretAccessKey("test-secret-access-key")
             .sessionToken("test-session-token")
             .build();
-    doReturn(credentials).when(awsCredentialVendor).vendAwsCredentials(any());
+    serverProperties.entrySet().stream()
+        .filter(e -> e.getKey().toString().contains("s3.bucketPath"))
+        .forEach(
+            e ->
+                doReturn(credentials)
+                    .when(awsCredentialVendor)
+                    .vendAwsCredentials(
+                        argThat(
+                            arg ->
+                                arg.getStorageScheme().equals("s3")
+                                    && arg.getStorageBase().contains(e.getValue().toString()))));
 
     azureCredentialVendor = mock(AzureCredentialVendor.class);
     AzureCredential azureCredential =
@@ -49,14 +69,96 @@ public abstract class BaseCRUDTestWithMockCredentials extends BaseCRUDTest {
             .sasToken("test-sas-token")
             .expirationTimeInEpochMillis(System.currentTimeMillis() + 6000)
             .build();
-    doReturn(azureCredential).when(azureCredentialVendor).vendAzureCredential(any());
+    serverProperties.entrySet().stream()
+        .filter(e -> e.getKey().toString().contains("adls.storageAccountName"))
+        .forEach(
+            e ->
+                doReturn(azureCredential)
+                    .when(azureCredentialVendor)
+                    .vendAzureCredential(
+                        argThat(
+                            arg ->
+                                arg.getStorageScheme().equals("abfs")
+                                    && arg.getStorageBase().contains(e.getValue().toString()))));
 
     gcpCredentialVendor = mock(GcpCredentialVendor.class);
     AccessToken accessToken =
         new AccessToken("test-token", Date.from(Instant.now().plusSeconds(10 * 60)));
-    doReturn(accessToken).when(gcpCredentialVendor).vendGcpToken(any());
+    serverProperties.entrySet().stream()
+        .filter(e -> e.getKey().toString().contains("gcs.bucketPath"))
+        .forEach(
+            e ->
+                doReturn(accessToken)
+                    .when(gcpCredentialVendor)
+                    .vendGcpToken(
+                        argThat(
+                            arg ->
+                                arg.getStorageScheme().equals("gs")
+                                    && arg.getStorageBase().contains(e.getValue().toString()))));
 
     credentialOperations =
         new CredentialOperations(awsCredentialVendor, azureCredentialVendor, gcpCredentialVendor);
+  }
+
+  protected String getTestCloudPath(String scheme, boolean isConfiguredPath) {
+    String bucket = isConfiguredPath ? "test-bucket0" : "test-bucket1";
+    return switch (scheme) {
+        case "s3" -> "s3://" + bucket + "/test";
+        case "abfs", "abfss" -> "abfs://test-container@" + bucket + ".dfs.core.windows.net/test";
+        case "gs" -> "gs://" + bucket + "/test";
+        default -> throw new IllegalArgumentException("Invalid scheme");
+    };
+  }
+
+  protected void assertTemporaryCredentials(
+      TemporaryCredentials temporaryCredentials, String scheme) {
+    switch (scheme) {
+      case "s3":
+        assertThat(temporaryCredentials.getAwsTempCredentials()).isNotNull();
+        assertThat(temporaryCredentials.getAwsTempCredentials().getSessionToken())
+            .isEqualTo("test-session-token");
+        assertThat(temporaryCredentials.getAwsTempCredentials().getAccessKeyId())
+            .isEqualTo("test-access-key-id");
+        assertThat(temporaryCredentials.getAwsTempCredentials().getSecretAccessKey())
+            .isEqualTo("test-secret-access-key");
+        break;
+      case "abfs":
+      case "abfss":
+        assertThat(temporaryCredentials.getAzureUserDelegationSas()).isNotNull();
+        assertThat(temporaryCredentials.getAzureUserDelegationSas().getSasToken())
+            .isEqualTo("test-sas-token");
+        break;
+      case "gs":
+        assertThat(temporaryCredentials.getGcpOauthToken()).isNotNull();
+        assertThat(temporaryCredentials.getGcpOauthToken().getOauthToken()).isEqualTo("test-token");
+        break;
+      default:
+        fail("Invalid scheme");
+        break;
+    }
+  }
+
+  protected void checkExceptionForScheme(Runnable functionCall, String scheme) {
+    if ("s3".equals(scheme)) {
+      assertThatThrownBy(functionCall::run, "Expected S3Exception for scheme: " + scheme)
+          .isInstanceOf(StsException.class);
+    } else if ("abfs".equals(scheme)) {
+      assertThatThrownBy(functionCall::run, "Expected AbfsException for scheme: " + scheme)
+          .isInstanceOf(CompletionException.class);
+    } else if ("gs".equals(scheme)) {
+      assertThatThrownBy(functionCall::run, "Expected GSException for scheme: " + scheme)
+          .isInstanceOf(BaseException.class);
+    } else {
+      throw new IllegalArgumentException("Unsupported scheme: " + scheme);
+    }
+  }
+  protected static Stream<Arguments> provideTestArguments() {
+    List<String> clouds = List.of("s3", "abfs", "gs");
+    List<Boolean> isConfiguredPathFlags = List.of(true, false);
+
+    // Cartesian product of clouds and isConfiguredPathFlags
+    return clouds.stream()
+            .flatMap(cloud -> isConfiguredPathFlags.stream()
+                    .map(isConfiguredPath -> Arguments.of(cloud, isConfiguredPath)));
   }
 }
